requirements.txt
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
PyYAML==3.12
Werkzeug==0.12.2
